Moved error functions upward
